+++ GraphemeSplitterJS/Character.GetGraphemeBreakPropertyV10.mjs
@@ -1,7 +1,6 @@
 //#if UnicodeVersionV10
 //generated by GraphemeBreakPropertyCodeGenerator
 // orginal https://github.com/ufcpp/GraphemeSplitter/blob/master/GraphemeSplitter/Character.GetGraphemeBreakPropertyV10.cs
-// js ver. https://github.com/foliojs/grapheme-breaker
 
 const getProperty = function(codePoint) {
     //        public static GraphemeBreakProperty GetGraphemeBreakProperty(uint codePoint)
