+++ GraphemeSplitterJSTest/test.mjs
@@ -1,4 +1,4 @@
-//import gsplit from 'https://taisukef.github.io/GraphemeSplitter/GraphemeSplitterJS/StringSplitter.mjs'
-import gsplit from '../GraphemeSplitterJS/StringSplitter.mjs'
+//import gsplit from 'https://taisukef.github.io/GraphemeSplitter/GraphemeSplitterJS/StringSplitter.Grapheme.mjs'
+import gsplit from '../GraphemeSplitterJS/StringSplitter.Grapheme.mjs'
 
 console.log(gsplit.split("👨‍👨‍👧‍👦👩‍👩‍👧‍👦👨‍👨‍👧‍👦"))
